+++ src/transport/irc/irc_lib_sup.erl
@@ -8,7 +8,7 @@
 -behaviour(supervisor).
 
 %% API
--export([start_link/0, start_irc_client/6]).
+-export([start_link/0, start_irc_client/7]).
 
 %% Supervisor callbacks
 -export([init/1]).
@@ -28,16 +28,17 @@ start_link() ->
                        Port :: integer(),
                        Channel :: {binary(), binary()}, 
                        Nick :: binary(),
-                       UseSsl :: boolean()) 
+                       UseSsl :: boolean(),
+                       ReconnectTimeout :: integer()) 
                        -> {ok, Pid :: pid()} | {error, Reason :: term()}.
-start_irc_client(CallbackModule, Host, Port, Channel, Nick, UseSsl) ->
+start_irc_client(CallbackModule, Host, Port, Channel, Nick, UseSsl, ReconnectTimeout) ->
     % Check use ssl or not
     SocketMod = case UseSsl of
       true -> ssl;
       false -> gen_tcp
     end,
     Child = {irc_lib_client, 
-                {irc_lib_client, start_link, [CallbackModule, Host, Port, SocketMod, Channel, Nick]},
+                {irc_lib_client, start_link, [CallbackModule, Host, Port, SocketMod, Channel, Nick, ReconnectTimeout]},
                  temporary, 2000, worker, []
              },
     % run new irc client